LPD-89874 refactor: move classes implementations of langchain4j classes to langchain4j directory
