+++ modules/dxp/apps/ai-hub/ai-hub-impl/src/main/java/com/liferay/ai/hub/internal/langchain4j/observability/api/listener/InputGuardrailExecutedListenerImpl.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.ai.hub.internal.guardrail.listener;
+package com.liferay.ai.hub.internal.langchain4j.observability.api.listener;
 
 import com.liferay.portal.workflow.kaleo.runtime.ExecutionContext;
